backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
pymysql==1.0.2
python-dotenv==0.19.1
sqlalchemy==1.4.23
werkzeug==2.0.1 